+++ core/netutil/tcp_soreuse_control_unix.go
@@ -1,3 +1,4 @@
+//go:build !windows && !wasm
 // +build !windows,!wasm
 
 package netutil